requirements.txt
annotated-types==0.7.0
anyio==4.9.0
aws-wsgi==0.2.7
black==25.1.0
blinker==1.9.0
boto3==1.37.28
botocore==1.37.28
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
google-auth==2.38.0
google-genai==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pulumi-esc-sdk==0.12.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pydantic==2.11.2
pydantic_core==2.33.1
pyflakes==3.3.2
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.0.7
websockets==15.0.1
Werkzeug==3.1.3
